Move "broadcast enabled" controlobject to BroadcastManager
